requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
docker==5.0.3
idna==3.3
pytz==2022.1
pywin32==227
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
websocket-client==1.3.2
